requirements.txt
tensorflow_gpu==1.9.0
tensorflow-probability==0.7.0
tensorboard==1.9.0
tensorflow-gan==1.0.0.dev0
tensorflow-metadata==0.14.0
tensorflow-datasets==1.2.0
mesh-tensorflow==0.0.5
tensor2tensor==1.14.0
matplotlib
librosa
tqdm
jiwer
pandas
numpy
textgrid
praat-parselmouth
soundfile
pathos
azure-cognitiveservices-speech
glob2